Enable gzip compression for both API and SPA
This is important for serving the SPA assets, as they can be pretty
large (>= 0.5 mb), but may also be useful for retrieving list of
snippets from the API as well.

Compression ratio and minimum size values are chosen, so that we try
to find a good balance between the length of the content sent over
the wire and the CPU load.