Add support for HTTPS in webserver config
Allow for configuring nginx to listen for HTTPS connections, if SSL
certificates are provided. The certificates and corresponding private
keys must be installed on the target host manually before executing
this playbook. Paths to these files are specified by the means of
the following variables:

  xsnippet_api_ssl_cert
  xsnippet_api_ssl_key
  xsnippet_spa_ssl_cert
  xsnippet_spa_ssl_key

It's also possible to configure a `307 Temporary Redirect` from HTTP
to HTTPS by the means of the following variables:

  xsnippet_api_https_redirect
  xsnippet_spa_https_redirect

Note, that we use a temporary redirect instead of a permanent one,
so that web browsers do not cache it and we can tweak things, if
needed, without having to tell people to clear their caches.

API does not force a redirect to HTTPS for now, as we want to allow
for local development of xsnippet-web.

Closes #13