Send all Docker container logs to journald
Make sure we have all the logs stored in the same place. This way
they are easier to manage, e.g. if we want to forward them somewhere
or simply configure periodic backups.

Update access log formats of xsnippet-api and xsnippet-web-backend,
so that we do not have duplicate timestamp values in logs (as
journald will add its own timestamp to every log entry).

One can get the logs by the means of journalctl, e.g.:

    $ journalctl -f CONTAINER_TAG=xsnippet_api

Note, that we assign a unique tag for each service in a Docker Swarm
stack, as names/ids of containers change on each update and we can't
rely on them.