Revise Cross-Origin Resource Sharing configuration
Cross-Origin Resource Sharing (CORS) is a mechanism that uses additional
HTTP headers to tell a browser to let a web application running at one
origin (domain) have permission to access selected resources from a
server at a different origin. A web application makes a cross-origin
HTTP request when it requests a resource that has a different origin
(domain, protocol, and port) than its own origin.

Since XSnippet API runs on api.xsnippet.org, and XSnippet Web
xsnippet.org, we, by definition, require CORS to let the latter to make
requests to the former.

Previously we had a copy pasted version of CORS configuration for Nginx,
not diving into the details of what's required and what's not. Now we
finally shaped it to our needs and specified only what's required.