Land IaaC configuration
Previously we used to have production configuration defined in CD job,
and production defaults defined in roles themselves. This wasn't great
because roles are designed to be agnostic and moveable between systems,
so they must have defaults that would make sense to everybody.

This patch lays down IaaC foundations. In particular it does the
following things:

 * Adds 'production' and 'ci' inventories on disk.
 * Sets variables per inventories without hardcoding them in CI/CD jobs.
 * Moves domains from 'xsnippet' role defaults to 'production'
   inventory.