Ensure variables are prefixed with the role name
The best practice is using name of the role as the prefix for variables.
The reason why we do this could be the situation when we need to place
variables for different roles in one place. Like playbook which calls
multiple roles. What will happen if all roles have variable port for
instance? By adding prefix we ensure that variables will not be
overwritten by other variables.