Do not resolve letsencrypt SSL cert/pkey
Letsencrypt saves newly generated SSL cert/pkey-s under new names, and
updates existing symlinks to point to the new keys. That means if we
resolve symlinks in Ansible playbook, the keys that are mounted inside
Nginx container won't be updated and will always be old.

This patch fixes the issue by mounting the whole letsencrypt directory
into Nginx container and by pointing Nginx to the symlinks.