Create backups of PostgreSQL databases
Adds a new systemd timer/service pair that triggers a plain-text SQL
dump (compressed with gzip) of a PostgreSQL database if backup schedule
is set in the corresponding `postgres_users` entry. The backups are
stored in `/var/lib/postgresql-backups` and are not uploaded anywhere
else yet (that will be done separately in a different role).

We could put this logic into a new role, but it seemed like a natural
extension of the `postgres_users` configuration which already lives in
the `postgres` role.