Add a primitive backup rotation mechanism
We have daily backups that periodically consume all available disk
space. Currently, I clean up those files once in awhile using a Rust
program I wrote back in the day, but I would like to make this part
of a normal setup. We _could_ use that Rust program, but it gets a
bit annoying with building and shipping the binary. For something
simple like this Python seems to be a better fit.

This updates the existing systemd unit to run the rotation script
after each successful backup. Currently, in the dry run mode. Outside
of dry run, the script will keep 30 most recent backups and delete the
rest.