Use infinite restarts for xsnippet-api.service
systemd defaults are fairly strict (<= 5 restarts within 10s) before
it gives up on the service. And we would prefer the service to recover
automatically after transient failures even if takes longer than that.

E.g. yesterday, I found that xsnippet-api had been down after the last
machine reboot because it could not fetch the Auth0 key on start, but
the same thing could happen with database connection as well.