Fix apt-key deprecation warning
apt-key is deprecated and is going to be removed in the next Debian
stable release. The suggested migration path is to use the so called
DEB822 source format. Among other things, this new format allows to
specify the package signing key. Conveniently, Ansible provides a module
that supports DEB822-style source definitions.

Configured keys are fetched and stashed in /etc/apt/trusted.gpg.d,
similarly to how this was handled with apt-key. The crucial difference
is that the scope of each key is limited to a single source now.